feat 💥: Move tasks out of ext namespace
