+++ dis_snek/models/snek/__init__.py
@@ -9,3 +9,4 @@ from .iterator import *
 from .listener import *
 from .scale import *
 from .wait import *
+from .tasks import *
