+++ dis_snek/models/snek/scale.py
@@ -6,7 +6,7 @@ from typing import Awaitable, List, TYPE_CHECKING, Callable, Coroutine
 import dis_snek.models.snek as snek
 from dis_snek.client.const import logger_name, MISSING
 from dis_snek.client.utils.misc_utils import wrap_partial
-from dis_snek.ext.tasks import Task
+from dis_snek.models.snek.tasks import Task
 
 if TYPE_CHECKING:
     from dis_snek.client import Snake
